テクスチャのvalidationを二次元配列を参照するように変更
