+++ .gitignore
@@ -27,3 +27,4 @@ compile_commands.json
 SUBJECT.md
 CODING_GUIDELINES.md
 .codex
+AGENTS.md
