+++ includes/cub3d.h
@@ -6,15 +6,16 @@
 /*   By: tshimizu <tshimizu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/06 22:16:04 by tshimizu          #+#    #+#             */
-/*   Updated: 2026/04/12 14:51:37 by tshimizu         ###   ########.fr       */
+/*   Updated: 2026/04/13 17:53:50 by shunwata         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef CUB3D_H
 # define CUB3D_H
 
-# include "read.h"
-# include "utils.h"
 # include "validation.h"
+# include "read.h"
+# include "parse.h"
+# include "utils.h"
 
 #endif // CUB3D_H
