+++ includes/parse.h
@@ -1,23 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   read.h                                             :+:      :+:    :+:   */
+/*   parse.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tshimizu <tshimizu@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: shunwata <shunwata@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/04/06 22:16:04 by tshimizu          #+#    #+#             */
-/*   Updated: 2026/04/13 17:54:07 by shunwata         ###   ########.fr       */
+/*   Created: 2026/04/13 17:47:47 by shunwata          #+#    #+#             */
+/*   Updated: 2026/04/13 17:59:26 by shunwata         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef READ_H
-# define READ_H
+#ifndef PARSE_H
+# define PARSE_H
 
-# include "../libs/get_next_line/srcs/get_next_line.h"
-# include "../libs/libft/libft.h"
-# include "utils.h"
 # include <stdbool.h>
 
-char	**read_file(int fd);
+bool	parse(char **lines);
 
-#endif // READ_H
+#endif // PARSE_H
