+++ includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: tshimizu <tshimizu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/06 22:16:04 by tshimizu          #+#    #+#             */
-/*   Updated: 2026/04/12 14:51:29 by tshimizu         ###   ########.fr       */
+/*   Updated: 2026/04/13 17:52:46 by shunwata         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,5 +17,8 @@
 
 void	free_array(char **arr);
 void	ft_noop(void *ptr);
+
+bool	is_space(char c);
+char	*skip_space(const char *line);
 
 #endif // UTILS_H
